backend/requirements.txt
# MedVoice AI Backend Requirements
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0

# ASR
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0

# LLM
mistralai>=0.1.0

# Audio
pydub>=0.25.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
